Refactor the quote data logic in index.ts to a utils helper method for the quote data
