+++ api/index.ts
@@ -1,7 +1,6 @@
 import type { VercelRequest, VercelResponse } from '@vercel/node';
 import { renderQuoteCard } from '../src/renderQuoteCard';
-import { QuoteData } from '../src/templates/QuoteCard';
-import { fetchQuote } from '../src/utils/fetchQuote';
+import { formatQuoteData, QuoteData } from '../src/utils/formatQuoteData';
 
 interface Query {
   quote: string;
@@ -11,21 +10,7 @@ interface Query {
 const handler = async (request: VercelRequest, response: VercelResponse) => {
   const { quote, author } = request.query as unknown as Query;
 
-  let data: QuoteData;
-
-  if (quote && author) {
-    data = {
-      quote,
-      author,
-    };
-  } else if (quote) {
-    data = {
-      quote,
-      author: 'Anonymous',
-    };
-  } else {
-    data = fetchQuote();
-  }
+  const data: QuoteData = formatQuoteData({ quote, author });
 
   response.setHeader('Content-Type', 'image/svg+xml');
   response.setHeader('Cache-Control', `public, max-age=300`);
