+++ src/renderQuoteCard.ts
@@ -1,4 +1,5 @@
-import { QuoteCard, QuoteData } from './templates/QuoteCard';
+import { QuoteCard } from './templates/QuoteCard';
+import { QuoteData } from './utils/formatQuoteData';
 
 const renderQuoteCard = (data: QuoteData): string => {
   return QuoteCard.build(data);
