+++ src/utils/fetchQuote.ts
@@ -1,9 +1,9 @@
-import { QuoteData } from '../templates/QuoteCard';
+import { QuoteData } from '../utils/formatQuoteData';
 
 const fetchQuote = (): QuoteData => {
   const quote =
-    'The happiness of your life depends upon the quality of your thoughts.';
-  const author = 'Marcus Aurelius';
+    'If a man knows not to which port he sails, no wind is favorable.';
+  const author = 'Seneca';
 
   return { quote, author };
 };
